requirements-dev.txt
jj==2.7.2
vedro-jj==0.1.1
jj_district42==1.0.0
flake8==6.0.0
isort==5.12.0
mypy==1.3.0
mypy-extensions==1.0.0
